[FIX] xlsx: remove useless normalization
The borders are already normalized at export time. The code was
denormalizing the renormalizing it without reason.

Task: 3733743
X-original-commit: 0e0f31a38d77354f27744bb60dddf53479dce72b
