+++ src/types/xlsx.ts
@@ -211,7 +211,7 @@ export interface XLSXStyle {
 export interface ExtractedStyle {
   font: XLSXFont;
   fill: XLSXFill;
-  border: XLSXBorder;
+  border: number;
   numFmt: XLSXNumFormat | undefined;
   alignment: XLSXCellAlignment;
 }
