+++ src/xlsx/helpers/xml_helpers.ts
@@ -1,6 +1,14 @@
 import { DEFAULT_FONT_SIZE } from "../../constants";
 import { concat } from "../../helpers";
-import { XLSXStructure, XMLAttributes, XMLAttributeValue, XMLString } from "../../types/xlsx";
+import { BorderDescr, ExcelWorkbookData } from "../../types";
+import {
+  XLSXBorder,
+  XLSXBorderDescr,
+  XLSXStructure,
+  XMLAttributeValue,
+  XMLAttributes,
+  XMLString,
+} from "../../types/xlsx";
 import { XLSXExportXMLFile } from "./../../types/xlsx";
 
 // -------------------------------------
@@ -55,7 +63,26 @@ export function parseXML(
   return document;
 }
 
-export function getDefaultXLSXStructure(): XLSXStructure {
+function convertBorderDescr(descr: BorderDescr | undefined): XLSXBorderDescr | undefined {
+  if (!descr) {
+    return undefined;
+  }
+  return {
+    style: descr.style,
+    color: { rgb: descr.color },
+  };
+}
+
+export function getDefaultXLSXStructure(data: ExcelWorkbookData): XLSXStructure {
+  const xlsxBorders: XLSXBorder[] = Object.values(data.borders).map((border) => {
+    return {
+      left: convertBorderDescr(border.left),
+      right: convertBorderDescr(border.right),
+      bottom: convertBorderDescr(border.bottom),
+      top: convertBorderDescr(border.top),
+    };
+  });
+  const borders = [{}, ...xlsxBorders];
   return {
     relsFiles: [],
     sharedStrings: [],
@@ -78,7 +105,7 @@ export function getDefaultXLSXStructure(): XLSXStructure {
       },
     ],
     fills: [{ reservedAttribute: "none" }, { reservedAttribute: "gray125" }],
-    borders: [{}],
+    borders,
     numFmts: [],
     dxfs: [],
   };
